requirements.txt
elasticsearch~=7.11.0
pandas~=1.2.3
python-dotenv